Introduce PointerEvent.deviceProperties
This change replaces deviceId on the PointerEvent interface with a new interface, deviceProperties. DeviceProperties contains one member, uniqueId, which functionally behaves the same as the outgoing deviceId.

Spec: https://github.com/w3c/pointerevents/pull/495

This change also removes the setting of deviceId when creating a PointerCancel event, as there is no current application for device/unique id for a PointerCancel.

Bug: 330760871
Change-Id: I0f1a9f7d5589f790d94f498a38bfdf55b6f51073