Rewrite url-processing-invalid-base.svg to use an invalid base URL
This test seems to depend on non-standard URL behavior of Chromium
implicitly. A relative URL fragment like "#fragment" combined with the
*valid* base URL "invalid:" (please don't be confused, "invalid:" is a
valid URL in the sense of the URL Standard), results in an invalid URL.

In other words, Chromium throws an exception for:

> new URL("#fragment", "invalid:")

However, this is not a standard compliant behavior. The result should be
a valid URL, "invalid:#fragment".

Note: 'valid' (or 'invalid' ) in the context of this CL's description
means what the URL Standard defines here:

https://url.spec.whatwg.org/#valid-url-string

Roughly speaking, if URL constructor doesn't throw an exception, the URL
is valid. If URL constructor throws an exception, the URL is invalid.

e.g.
- new URL("invalid:") doesn't throw an exception, thus "invalid:" is a
  valid URL.
- new URL("invalid") throw an exception, thus "invalid" is NOT a valid
  URL.

This test is currently passing on Chromium, but failing on Safari and
Firefox because Chromium's current URL behavior is non compliant as
mentioned above.

https://wpt.fyi/results/svg/linking/reftests/url-processing-invalid-base.svg?label=experimental&label=master&aligned

If Chromium supports non-special URLs correctly
(https://crbug.com/40063064), then, `new URL("#fragment", "invalid:")'
becomes "invalid:#fragment" correctly. Thus, this test will start to
fail on Chromium too. That's a correct behavior and the test should be
updated.

One way to fix and preserve the test is to change the base url from
"invalid:" (this is a valid base URL), to "invalid" (invalid URL; no
colon at the end), and to swap "orange" and "green".

Then, `new URL("#fragment", "invalid") becomes invalid on all browsers,
which makes the test pass on all browsers. Additionally, and
importantly, the test will also pass in Chromium even if the flag for
https://crbug.com/40063064 is enabled.

This test was added in https://crrev.com/c/2115650.

Bug: 40621475,40063064
Change-Id: I46dee043ab02e7899a512e5a2cee3fc773848e5c