Fix popover-light-dismiss.html
We are failing the last subtest in popover-light-dismiss which clicks
and drags from the main frame to an iframe because chromedriver is
telling chrome that we are doing pointerUp in both the button in the
main frame and the body of the iframe at the same time. I'm not sure
why, but combining the actions sequences and using the main frame's
document as the origin for the pointerUp fixes this.

The failure only occurs with run_wpt_tests.py, not run_web_tests.py.

Bug: 325981758
Change-Id: I992a01a6baf78a7701da3dd8df76788c93ed0aaa