[anchor] Sort candidate list before handing off to OOFLayoutPart.
The order of the candidates now is important, when it wasn't
previously. The order is only important when there is a child that
has an anchor within it, so to mitigate the performance, only sort
when that is true.

The order can get inconsistent for a variety of reasons, the testcase
is when we insert them in the incorrect order during simplified layout
(as we don't keep the OOFs in the same list as the child fragments).

Fixed: 330756838
Change-Id: I7e4e08d104c5a75e96b6250ba234183ca3ab458f