[line-clamp] Hide out of flow boxes after the clamp point
This patch makes out of flow boxes that appear in the layout tree
after the clamp point (that is, in tree order after the end of the
last shown line) hidden for paint. This happens depending entirely on
the layout tree order, regardless of an absolute positioned box's
position or of the extent of a float.

This is achieved by storing a flag in `UnpositionedFloat` and in the
various `*OofPositionedNode` classes to mark the node as hidden for
paint. This flag is then checked when creating the constraint space
for the layout of the OOF box.

Bug: 40336192
Change-Id: Ia42751b2fa7e3fee89648e26600035241bda6299