Add web platform test for additional bids with negative targeting.
This is the second is a series of changes adding WPTs for additional
bids. This adds the basic infrastructure needed to join negative
interest groups, and sign and negatively target additional bids.
This also adds tests that verify the behavior of a correctly negatively
targeted additional bid as well as those that verify the behavior of
errors one might encounter while attempting to negatively target an
additional bid.

Bug: 40275797
Change-Id: Ib05c7e1340c25bf3daa88a49154c393d5e12a468