Second attempt to de-flake sandbox top navigation WPTs.
Remote context executors tend to be flaky when dealing with either
cross-origin subframe navigations or nested subframes. This CL attempts
to fix the last 2 flaking sandbox top-level navigation WPTs by rewriting
them to not use remote executors where the flakes are most likely to
happen.

For sandbox-top-navigation-child-cross-origin-frame:
- The iframe created is navigated to a cross-origin document, so that
is replaced with a navigation to an HTML page rather than a remote
context.

For sandbox-top-navigation-grandchild-frame-allow-top:
- The grandchild iframe being navigated to a remote context is the most likely culprit, so that is replaced with a navigation to an HTML page.

Bug: 329153138
Bug: 328860404
Change-Id: Ibcbf88b474ef33d04693f207afa3f1859dcc4d23